requirements.txt
asgiref==3.6.0
astroid==2.14.1
colorama==0.4.6
crispy-bootstrap5==0.7
dill==0.3.6
distlib==0.3.6
Django==4.1.5
django-crispy-forms==1.14.0
django-model-utils==4.3.1
django-notifications-hq==1.7.0
filelock==3.9.0
isort==5.12.0
jsonfield==3.1.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
Pillow==9.4.0
platformdirs==2.6.2
pylint==2.16.0
python-dotenv==0.21.1
pytz==2022.7.1
sqlparse==0.4.3
swapper==1.3.0
tomlkit==0.11.6
tzdata==2022.7
virtualenv==20.19.0
virtualenvwrapper-win==1.2.7
wrapt==1.14.1
